fix: quick fix for crash on deprecated setting
The max_accel_to_decel printer setting is deprecated according
to https://www.klipper3d.org/Config_Reference.html#printer.
Since the cast to float is unchecked, there is a crash if using
recent klipper when trying to create the cell for this setting.

The rows created should probably be dynamic (like temperature?)
but the blank row does no real harm and I thought a quick fix
might be okay for now.